kernel: new ring0 image

+++ ws/server/line.h
@@ -37,7 +37,7 @@ lineBackbufferDrawHorizontalLine (
     unsigned long x1,
     unsigned long y, 
     unsigned long x2,  
-    unsigned long color );
+    unsigned int color );
